requirements-dev.txt
backports.tarfile==1.2.0
black==25.1.0
build==1.3.0
commitizen==4.8.3
inflect==7.5.0
ipykernel==6.30.1
jaraco-collections==5.2.1
jaraco-functools==4.3.0
pip-chill==1.0.3
pip-upgrade-outdated==1.5
pip3-autoremove==2.0.0
pkginfo==1.12.1.2
pre-commit==4.3.0
pytest-cov==6.2.1
tomli==2.2.1
twine==6.1.0
# appnope==0.1.4 # Installed as dependency for ipykernel
# argcomplete==3.6.2 # Installed as dependency for commitizen
# asttokens==3.0.0 # Installed as dependency for stack-data
# autocommand==2.2.2 # Installed as dependency for jaraco.text
# certifi==2025.8.3 # Installed as dependency for requests
# cfgv==3.4.0 # Installed as dependency for pre-commit
# charset-normalizer==3.4.3 # Installed as dependency for commitizen, requests
# click==8.2.1 # Installed as dependency for black
# colorama==0.4.6 # Installed as dependency for commitizen
# comm==0.2.3 # Installed as dependency for ipykernel
# coverage==7.10.5 # Installed as dependency for pytest-cov
# debugpy==1.8.16 # Installed as dependency for ipykernel
# decli==0.6.3 # Installed as dependency for commitizen
# decorator==5.2.1 # Installed as dependency for ipython
# distlib==0.4.0 # Installed as dependency for virtualenv
# docutils==0.22 # Installed as dependency for readme-renderer
# executing==2.2.0 # Installed as dependency for stack-data
# filelock==3.19.1 # Installed as dependency for virtualenv
# id==1.5.0 # Installed as dependency for twine
# identify==2.6.13 # Installed as dependency for pre-commit
# idna==3.10 # Installed as dependency for requests
# importlib-metadata==8.7.0 # Installed as dependency for commitizen
# iniconfig==2.1.0 # Installed as dependency for pytest
# ipython==9.4.0 # Installed as dependency for ipykernel
# ipython-pygments-lexers==1.1.1 # Installed as dependency for ipython
# jaraco.classes==3.4.0 # Installed as dependency for keyring
# jaraco.context==6.0.1 # Installed as dependency for jaraco.text, keyring
# jaraco.functools==4.0.1 # Installed as dependency for jaraco.text, keyring
# jaraco.text==4.0.0 # Installed as dependency for jaraco-collections, jaraco.collections
# jedi==0.19.2 # Installed as dependency for ipython
# jinja2==3.1.6 # Installed as dependency for commitizen
# jupyter-client==8.6.3 # Installed as dependency for ipykernel
# jupyter-core==5.8.1 # Installed as dependency for ipykernel, jupyter-client
# keyring==25.6.0 # Installed as dependency for twine
# markdown-it-py==4.0.0 # Installed as dependency for rich
# markupsafe==3.0.2 # Installed as dependency for jinja2
# matplotlib-inline==0.1.7 # Installed as dependency for ipykernel, ipython
# mdurl==0.1.2 # Installed as dependency for markdown-it-py
# more-itertools==10.7.0 # Installed as dependency for inflect, jaraco-functools, jaraco.classes, jaraco.functools, jaraco.text
# mypy-extensions==1.1.0 # Installed as dependency for black
# nest-asyncio==1.6.0 # Installed as dependency for ipykernel
# nh3==0.3.0 # Installed as dependency for readme-renderer
# nodeenv==1.9.1 # Installed as dependency for pre-commit
# packaging==25.0 # Installed as dependency for black, build, commitizen, ipykernel, pip3-autoremove, pytest, twine
# parso==0.8.5 # Installed as dependency for jedi
# pathspec==0.12.1 # Installed as dependency for black
# pexpect==4.9.0 # Installed as dependency for ipython
# platformdirs==4.3.8 # Installed as dependency for black, jupyter-core, virtualenv
# pluggy==1.6.0 # Installed as dependency for pytest, pytest-cov
# prompt-toolkit==3.0.51 # Installed as dependency for ipython, questionary
# psutil==7.0.0 # Installed as dependency for ipykernel
# ptyprocess==0.7.0 # Installed as dependency for pexpect
# pure-eval==0.2.3 # Installed as dependency for stack-data
# pygments==2.19.2 # Installed as dependency for ipython, ipython-pygments-lexers, pytest, readme-renderer, rich
# pyproject-hooks==1.2.0 # Installed as dependency for build
# pytest==8.4.1 # Installed as dependency for pytest-cov
# python-dateutil==2.9.0.post0 # Installed as dependency for jupyter-client
# pyyaml==6.0.2 # Installed as dependency for commitizen, pre-commit
# pyzmq==27.0.2 # Installed as dependency for ipykernel, jupyter-client
# questionary==2.1.0 # Installed as dependency for commitizen
# readme-renderer==44.0 # Installed as dependency for twine
# requests==2.32.5 # Installed as dependency for id, requests-toolbelt, twine
# requests-toolbelt==1.0.0 # Installed as dependency for twine
# rfc3986==2.0.0 # Installed as dependency for twine
# rich==14.1.0 # Installed as dependency for twine
# six==1.17.0 # Installed as dependency for python-dateutil
# stack-data==0.6.3 # Installed as dependency for ipython
# tomlkit==0.13.3 # Installed as dependency for commitizen
# tornado==6.5.2 # Installed as dependency for ipykernel, jupyter-client
# traitlets==5.14.3 # Installed as dependency for ipykernel, ipython, jupyter-client, jupyter-core, matplotlib-inline
# typeguard==4.4.4 # Installed as dependency for inflect
# typing-extensions==4.15.0 # Installed as dependency for typeguard
# urllib3==2.5.0 # Installed as dependency for requests, twine
# virtualenv==20.34.0 # Installed as dependency for pre-commit
# wcwidth==0.2.13 # Installed as dependency for prompt-toolkit
# zipp==3.23.0 # Installed as dependency for importlib-metadata
